feat: add createProduct lambda-fn and errorHandler function
