+++ product-service/lambda-functions/errorHandler.ts
@@ -33,7 +33,7 @@ function createErrorResponse(
     statusCode,
     headers: {
       "Access-Control-Allow-Origin": "*",
-      "Access-Control-Allow-Methods": "GET",
+      "Access-Control-Allow-Methods": "OPTIONS,GET,POST,PUT,DELETE",
       "Access-Control-Allow-Headers": "Content-Type",
       "Content-Type": "application/json",
     },
