--- product-service/lambda-functions/handlers/createProduct.ts
@@ -1,34 +1,0 @@
-import * as AWS from "aws-sdk";
-import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
-import { handleAPIGatewayError } from "./errorHandler";
-import { ProductInfo } from "../types/product.interface";
-
-const dynamodb = new AWS.DynamoDB.DocumentClient();
-const tableName = "ProductsTable";
-
-export const handler = async (
-  event: APIGatewayProxyEvent,
-): Promise<APIGatewayProxyResult> => {
-  try {
-    const body = JSON.parse(event.body || "{}") as ProductInfo;
-    const { title, description, price } = body;
-    const params: AWS.DynamoDB.DocumentClient.PutItemInput = {
-      TableName: tableName,
-      Item: { title, description, price },
-    };
-
-    await dynamodb.put(params).promise();
-    return {
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin": "*",
-        "Access-Control-Allow-Methods": "PUT",
-        "Access-Control-Allow-Headers": "Content-Type",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ message: "Product created successfully" }),
-    };
-  } catch (error) {
-    return handleAPIGatewayError(error);
-  }
-};
